% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{glance.ritest}
\alias{glance.ritest}
\title{Glance at an `ritest` object}
\usage{
\method{glance}{ritest}(x, ...)
}
\arguments{
\item{x}{An object produced by the `ritest` function.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A data frame of goodness-of-fit statistics that conforms to the
  `broom` package specifications.
}
\description{
Glance at an `ritest` object
}
\examples{
est = lm(yield ~ N + P + K, data = npk)
est_ri = ritest(est, 'N', reps = 1e3, seed = 1234L, pcores = 2L)
tidy(est_ri)
glance(est_ri)
}
