% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritest.R
\name{prep_split_var}
\alias{prep_split_var}
\title{Prep split variables}
\usage{
prep_split_var(x, ...)
}
\arguments{
\item{x}{String or one-sided formula.}

\item{...}{Additional helpers objects, e.g. from a model.matrix.}
}
\description{
Internal function for prepping the split vars, i.e. strata
  and/or clusters. Will extract and assign the underlying data (used in the
  model object) to the parent environment if required.
}
