% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritest.R
\name{print.ritest}
\alias{print.ritest}
\title{A print method for ritest objects}
\usage{
\method{print}{ritest}(x, ...)
}
\arguments{
\item{x}{An ritest object.}

\item{...}{Currently ignored.}
}
\description{
Printed display of ritest objects. Tries to mimic the display
  of the equivalent Stata routine.
}
