% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritest.R
\name{plot.ritest}
\alias{plot.ritest}
\title{A plot method for ritest objects}
\usage{
\method{plot}{ritest}(
  x,
  type = c("density", "hist"),
  highlight = c("lines", "fill", "both", "none"),
  highlight_par = FALSE,
  breaks = "auto",
  family = NULL,
  ...
)
}
\arguments{
\item{x}{An ritest object.}

\item{type}{Character. What type of plot do you want?}

\item{highlight}{Character. How do you want to highlight the H0 rejection
regions in the distribution tails?}

\item{highlight_par}{Logical. Should we highlight the parametric H0 rejection
regions too?}

\item{breaks}{Character. Histogram plot only. What type of breaks do you
want? The default method creates more breaks than the standard R behaviour.
You can revert to the latter by selecting NULL.}

\item{family}{Character. The font family. Defaults to 'HersheySans' instead
of R's normal Arial plotting font.}

\item{...}{Other plot arguments. Currently ignored.}
}
\description{
Nice plots of your ritest objects.
}
