% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritest.R
\name{ritest}
\alias{ritest}
\title{Perform random inference on a model object}
\usage{
ritest(
  resampvar,
  object,
  reps = 100,
  pvals = c("both", "left", "right"),
  strata = NULL,
  cluster = NULL,
  level = 0.95,
  seed = NULL,
  ...
)
}
\arguments{
\item{resampvar}{Character. The name of the variable (coefficient) that you want
to test. At present, only a single variable is permitted.}

\item{object}{Model object containing the `resampvar` variable. At present,
only `stats::lm` and `fixest::feols` models are supported.}

\item{reps}{Integer. The number of repetitions (permutations) in the RI
simulation. Default is 100, but you probably want more that that. (Alwyn
Young has suggested at least 2000 in a research setting.)}

\item{pvals}{Character. How should the p-values should be computed? The
default is "both", which means that two-sided p-values will be computed.
Alternatively, users can specify one-sided p-values with either "left" or
"right".}

\item{strata}{Character or one-sided formula. Permute `resampvar` with strata
strata? See Details and Examples below.}

\item{cluster}{Character or one-sided formula. Keep `resampvar` constant with
clusters? See Details and Examples below.}

\item{level}{Numeric. The desired confidence level. Default if 0.95.}

\item{seed}{Integer. Random seed for reproducible results.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
An list object of class `ritest`. Default print and plotting methods
  are supported.
}
\description{
Perform random inference (RI) testing on a model object, e.g. a
  coefficient from a linear regression model. It tries to mimic the `ritest`
  Stata routine (Hess, 2017) in its design and functionality. Only a subset
  of this functionality is currently supported. However, it does appear to be
  a lot faster.
}
\details{
This function is experimental and functionality is still quite
  limited. Albeit, that it does support the most likely use case for RI on a
  regression model, i.e. permutation testing of a coefficient value. Present
  limitations include: only `lm` and `fixest::feols` model objects are
  supported; only one permutation (RI) test is allowed; and only one strata
  and/or cluster variable, respectively, can be supplied. I hope to resolve
  these limitations as time permits.
}
\examples{
library(fixest)
(mod = feols(yield ~ N + P + K | block, data = npk))
mod_ri = ritest('N1', mod, strata = ~block) ## We need 'N1' b/c that's the factored name in the model
mod_ri
plot(mod_ri)
plot(mod_ri, type = 'hist', highlight = 'fill', highlight_par = TRUE)
}
\seealso{
[print.ritest()], [plot.ritest()]
}
