% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.ritest}
\alias{tidy.ritest}
\title{Tidy an `ritest` object}
\usage{
\method{tidy}{ritest}(x, conf.int = TRUE, ...)
}
\arguments{
\item{x}{An object produced by the `ritest` function.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A "tidy" `data.frame` of summary statistics that conforms to the
  `broom` package specification.
}
\description{
Tidy an `ritest` object
}
\examples{
est = lm(yield ~ N + P + K, data = npk)
est_ri = ritest(est, 'N', reps = 1e3, seed = 1234L, pcores = 2L)
tidy(est_ri)
glance(est_ri)
}
